split up antenna data to fix hot reloading
